Fixed condition in trigger script to not create a broken symlink when both packages are getting removed.
git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat/trunk@15885 4e558342-562e-0410-864c-e07659590f8c

+++ voms-api-java/osg/voms-api-java.spec
@@ -7,7 +7,7 @@
 
 Name:		voms-api-java
 Version:	2.0.8
-Release:	1.3%{?dist}
+Release:	1.4%{?dist}
 Summary:	Virtual Organization Membership Service Java API
 
 Group:		Development/Libraries
@@ -150,11 +150,14 @@ rm -rf $RPM_BUILD_ROOT
 # This results in the symlink getting removed if both packages are installed and vomsjapi is erased.
 # Recreate the symlink if that is the case.
 # Note that we do not want to create the symlink if this package is getting uninstalled too.
-if [ $1 -gt 0 ]; then
+if [[ $1 -gt 0 && -e %{_javadir}/%{name}.jar ]]; then
     ln -sf %{name}.jar %{_javadir}/vomsjapi.jar
 fi
 
 %changelog
+* Tue Nov 20 2012 Matyas Selmeci <matyas@cs.wisc.edu> - 2.0.8-1.4
+- Fixed condition in triggerscript to not create a broken package if both packages are being removed.
+
 * Tue Nov 20 2012 Matyas Selmeci <matyas@cs.wisc.edu> - 2.0.8-1.3
 - Add triggerscript to fix broken symlink issue with older vomsjapi package
